requirements.txt
Flask==2.3.3
httpx==0.27.0
flask-cors==4.0.0
cachetools==5.3.1
requests==2.32.3
pycryptodome==3.20.0
protobuf==4.25.3
google==3.0.0
Werkzeug==2.3.7